frontend: common: Refactoring common/ imports in frontend directory
