+++ frontend/src/components/resourceMap/KubeObjectGlance/DeploymentGlance.tsx
@@ -18,7 +18,7 @@ import { Box } from '@mui/system';
 import { useTranslation } from 'react-i18next';
 import { KubeCondition } from '../../../lib/k8s/cluster';
 import Deployment from '../../../lib/k8s/deployment';
-import { StatusLabel } from '../../common';
+import { StatusLabel } from '../../common/Label';
 
 export function DeploymentGlance({ deployment }: { deployment: Deployment }) {
   const { t } = useTranslation();
