+++ frontend/src/components/resourceMap/KubeObjectGlance/PodGlance.tsx
@@ -17,7 +17,7 @@
 import { Box } from '@mui/system';
 import { useTranslation } from 'react-i18next';
 import Pod from '../../../lib/k8s/pod';
-import { StatusLabel } from '../../common';
+import { StatusLabel } from '../../common/Label';
 import { makePodStatusLabel } from '../../pod/List';
 
 export function PodGlance({ pod }: { pod: Pod }) {
