+++ frontend/src/components/resourceMap/KubeObjectGlance/ReplicaSetGlance.tsx
@@ -17,7 +17,7 @@
 import { Box } from '@mui/system';
 import { useTranslation } from 'react-i18next';
 import ReplicaSet from '../../../lib/k8s/replicaSet';
-import { StatusLabel } from '../../common';
+import { StatusLabel } from '../../common/Label';
 
 export function ReplicaSetGlance({ set }: { set: ReplicaSet }) {
   const { t } = useTranslation();
