+++ frontend/src/components/resourceMap/details/GraphNodeDetails.tsx
@@ -17,7 +17,7 @@
 import Box from '@mui/material/Box';
 import Card from '@mui/material/Card';
 import { useTranslation } from 'react-i18next';
-import { ActionButton } from '../../common';
+import ActionButton from '../../common/ActionButton';
 import { GraphNode } from '../graph/graphModel';
 import { KubeObjectDetails } from './KubeNodeDetails';
